+++ src/printer/simplify.ml
@@ -138,7 +138,7 @@ let print_decl info fmt d = match d.d_node with
   | Dprop (Pgoal, pr, f) ->
       fprintf fmt "@[;; %a@]@\n" print_ident pr.pr_name;
       (match pr.pr_name.id_loc with
-        | Some loc -> fprintf fmt " @[;; %a@]@\n" Loc.gen_report_position loc
+        | Some loc -> fprintf fmt " @[;; File %a@]@\n" Loc.pp_position loc
         | None -> ());
       fprintf fmt "@[<hov 2>%a@]@\n" (print_fmla info) f
   | Dprop (Plemma, _, _) ->